% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr2sse-package.R
\docType{package}
\name{nlmixr2sse-package}
\alias{nlmixr2sse}
\alias{nlmixr2sse-package}
\title{nlmixr2sse package}
\description{
Stochastic simulation and estimation tools for nlmixr2 fits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kestrel99/nlmixr2sse}
  \item Report bugs at \url{https://github.com/kestrel99/nlmixr2sse/issues}
}

}
\author{
\strong{Maintainer}: Justin Wilkins \email{justin.wilkins@occams.com} (\href{https://orcid.org/0000-0002-7099-9396}{ORCID})

}
\keyword{internal}
