% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sse.R
\name{plotSSEDiagnostics}
\alias{plotSSEDiagnostics}
\title{Plot a multi-panel SSE diagnostics view}
\usage{
plotSSEDiagnostics(x, biasStatistic = "relative_bias", ...)
}
\arguments{
\item{x}{An \code{nlmixr2SSE} object.}

\item{biasStatistic}{Parameter-summary statistic to use in the bias panel.}

\item{...}{Reserved for future extensions.}
}
\value{
A patchwork object.
}
\description{
Plot a multi-panel SSE diagnostics view
}
