% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sse.R
\name{plotSSEOfvDistribution}
\alias{plotSSEOfvDistribution}
\title{Plot SSE OFV distributions}
\usage{
plotSSEOfvDistribution(x, models = NULL, style = c("ecdf", "histogram"), ...)
}
\arguments{
\item{x}{An \code{nlmixr2SSE} object.}

\item{models}{Optional model labels to retain.}

\item{style}{Plot style: empirical CDF or histogram.}

\item{...}{Reserved for future extensions.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot SSE OFV distributions
}
